Implemented feature of selecting multiple tags and add them together.
An addition problem fixed
Flaw: The tables of MenuScene are showing the top most folder whenever comeback from another scene.

Fix: make the current folder varible public and set it to be the root folder whenever the user select a new root folder,
instead of set it whenever a MenuScene is initialized.